+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.1.0
+- Added `multi` parameter to `SocketConnector.serverToSocket` - whether to
+  create new connections on the "B" side every time there is a new "A" side
+  connection to the bound server port. Also added `onConnect` parameter,
+  so that callers can be informed when every new connection is made, and
+  can thus take whatever action they require.
+
 ## 2.0.1
 - Removed an unnecessary dependency